move tools/ folder to internal/
