+++ internal/tools/cli/version.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/lightning-dabbler/sportscrape/tools/common"
+	"github.com/lightning-dabbler/sportscrape/internal/tools/common"
 	"github.com/lightning-dabbler/sportscrape/version"
 	"github.com/spf13/cobra"
 )
