+++ internal/tools/tools.go
@@ -1,5 +1,3 @@
-//go:build tools
-
 package tools
 
 import (
